+++ packages/terra-date-picker/CHANGELOG.md
@@ -11,7 +11,7 @@ Unreleased
 
 ### Changed
 * Update react/react-dom peer-dependency to latest stable minor release
-* Forked React Datepicker
+* Forked React Datepicker at version 0.64.0
 * Removed react-datepicker as dependency
 * Added react-popper and react-onclickoutside as dependencies